fixed syyntax errors

+++ lab2/src/main/java/com/SalesReducer.java
@@ -6,7 +6,7 @@ import org.apache.hadoop.mapreduce.*;
 import org.apache.hadoop.mapreduce.Reducer.*;
 
 public class SalesReducer
-extends Reducer<Text, IntWritable, Text, DoubleWritable> {
+extends Reducer<Text, IntWritable, Text, IntWritable> {
    @Override
    public void reduce(Text date,
    Iterable<IntWritable> sales, Context context)
